Remove unused parent variable
The `ParserBlock` parent variable passed down the parser block tree is not defined for the root element, which causes an exception in PHP 7 (it used to be that class type hints always accepted null values).

Rather than making this nullable, the current parser block implementations do not use parent information and it makes to more sense and safer for parser blocks to be context agnostic. As a result, this pull request remove the parent parser block from both the `parse` functions and the `_insertIntoTree` functions.